refactor(storage): Split varint.rs, float.rs, utils.rs, matcher.rs, and in_iterator.rs into directory modules
Phase 3 of #330. Split 5 large files into directory modules by
extracting tests into separate files, bringing all files under 400 lines.

- encoding/varint.rs (734 lines) → varint/{mod,encode,decode,tests}.rs
- encoding/float.rs (466 lines) → float/{mod,tests}.rs
- keys/utils.rs (533 lines) → utils/{mod,tests}.rs
- index/matcher.rs (509 lines) → matcher/{mod,tests}.rs
- index/in_iterator.rs (493 lines) → in_iterator/{mod,tests}.rs

Pure code movement — zero behavioral changes. All 365 storage tests pass.
