+++ crates/storage/src/encoding/varint/mod.rs
@@ -1,0 +1,10 @@
+//! Variable-length integer encoding (varint/uvarint)
+
+mod decode;
+mod encode;
+
+#[cfg(test)]
+mod tests;
+
+pub use decode::*;
+pub use encode::*;
